requirements.txt
attr==0.3.2
ConfigParser==7.2.0
cryptography==46.0.3
Cython==3.1.6
dl==0.1.0
docutils==0.22.2
fastapi==0.121.0
filelock==3.20.0
HTMLParser==0.0.2
ipython==8.18.0
ipywidgets==8.1.5
Jinja2==3.1.2
keyring==25.6.0
numpy==2.3.4
protobuf==6.33.0
psycopg2_binary==2.9.11
pydantic==2.12.3
pyOpenSSL==25.3.0
railroad==0.5.0
redis==7.0.1
Sphinx==8.2.3
thread==2.0.5
trove_classifiers==2025.9.11.17
urllib3_secure_extra==0.1.0
xmlrpclib==1.0.1
